backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
openai==1.10.0
chromadb==0.4.22
sentence-transformers==2.2.2

pillow==10.2.0
pytesseract==0.3.10

tree-sitter==0.23.2
tree-sitter-python==0.23.2
streamlit
streamlit-cropper

tenacity==8.2.3
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
